Convert users from model to service setup.
