+++ src/server/api.js
@@ -1,21 +1,9 @@
-import bodyParser from 'body-parser';
 import express from 'express';
 
 import {
   createCampaignRequestHandler,
-  getUser,
-  getUsers,
+  createUserRequestHandler,
 } from './controllers';
-
-function requiresAuth(callback) {
-  return (req, res) => {
-    if (req.store.currentUser == null) {
-      return res.status(403).send();
-    }
-
-    return callback(req, res);
-  };
-}
 
 export function createApiRequestHandler(config) {
   const app = express();
@@ -29,20 +17,7 @@ export function createApiRequestHandler(config) {
   });
 
   app.use('/campaigns', createCampaignRequestHandler(config));
-
-  return app;
-}
-
-export function createUserRequestHandler() {
-  const app = express();
-
-  app.use(bodyParser.json());
-
-  app.route('/')
-    .get(requiresAuth(getUsers));
-
-  app.route('/:userId')
-    .get(requiresAuth(getUser));
+  app.use('/users', createUserRequestHandler(config));
 
   return app;
 }
