+++ src/server/controllers/campaigns.js
@@ -2,26 +2,21 @@ import bodyParser from 'body-parser';
 import express from 'express';
 import mysql from 'mysql';
 
-import {
-  fetchUserById,
-  fetchUsersById,
-  sanitizeUser,
-} from '../models';
-import { CampaignService } from '../services';
+import { CampaignService, UserService } from '../services';
 import { formatValidationErrors } from './util';
 
 function createCampaign(req, res) {
-  const { campaignService } = req;
+  const { campaignService, userService } = req;
 
   return campaignService.validateCampaign(req.body)
     .catch(errors => res.status(400).json({
       errors: formatValidationErrors(errors),
     }))
     .then(campaign => campaignService.saveCampaign(campaign))
-    .then(campaign => fetchUsersById(req.connection, campaign.relationships.creator.data.id)
+    .then(campaign => userService.fetchUsersById(campaign.relationships.creator.data.id)
       .then(users => ({
         data: campaign,
-        included: users.map(sanitizeUser),
+        included: users.map(UserService.sanitizeUser),
       }))
     )
     .then(response => res.status(201).json(response))
@@ -48,7 +43,7 @@ function deleteCampaign(req, res) {
 }
 
 function getCampaign(req, res) {
-  const { campaignService } = req;
+  const { campaignService, userService } = req;
 
   return campaignService.validateCampaignId(req.params)
     .catch(errors => res.status(400).json({
@@ -63,8 +58,8 @@ function getCampaign(req, res) {
       return campaign;
     })
     .then(campaign => Promise.all([
-      fetchUserById(req.connection, campaign.relationships.creator.data.id),
-      campaignService.fetchUsersByCampaignId(campaign.id),
+      userService.fetchUserById(campaign.relationships.creator.data.id),
+      userService.fetchUsersByCampaignId(campaign.id),
     ])
       .then(([creator, users]) => {
         const data = {
@@ -91,7 +86,7 @@ function getCampaign(req, res) {
 
         return {
           data,
-          included: [creator].concat(users.players).concat(users.invited).map(sanitizeUser),
+          included: [creator].concat(users.players).concat(users.invited).map(UserService.sanitizeUser),
         };
       })
     )
@@ -130,6 +125,10 @@ export function createCampaignRequestHandler(config) {
     req.connection = mysql.createConnection(config.mysql);
     // eslint-disable-next-line no-param-reassign
     req.campaignService = new CampaignService({
+      connection: req.connection,
+    });
+    // eslint-disable-next-line no-param-reassign
+    req.userService = new UserService({
       connection: req.connection,
     });
