+++ src/server/controllers/users.js
@@ -1,12 +1,14 @@
-import {
-  fetchUserById,
-  fetchUsers,
-  sanitizeUser,
-} from '../models';
+import bodyParser from 'body-parser';
+import express from 'express';
+import mysql from 'mysql';
+
+import { UserService } from '../services';
 
 export function getUsers(req, res) {
-  return fetchUsers(req.connection)
-    .then(users => users.map(user => sanitizeUser(user)))
+  const { userService } = req;
+
+  return userService.fetchUsers()
+    .then(users => users.map(UserService.sanitizeUser))
     .then(users => res.status(200).json({
       data: users,
     }))
@@ -14,9 +16,36 @@ export function getUsers(req, res) {
 }
 
 export function getUser(req, res) {
-  return fetchUserById(req.connection, req.params.userId)
-    .then(user => sanitizeUser(user))
+  const { userService } = req;
+
+  return userService.fetchUserById(req.params.userId)
+    .then(user => UserService.sanitizeUser(user))
     .then(user => res.status(200).json({
       data: user,
     }));
+}
+
+export function createUserRequestHandler(config) {
+  const app = express();
+
+  app.use(bodyParser.json());
+
+  app.use((req, res, next) => {
+    // eslint-disable-next-line no-param-reassign
+    req.connection = mysql.createConnection(config.mysql);
+    // eslint-disable-next-line no-param-reassign
+    req.userService = new UserService({
+      connection: req.connection,
+    });
+
+    next();
+  });
+
+  app.route('/')
+    .get(getUsers);
+
+  app.route('/:userId')
+    .get(getUser);
+
+  return app;
 }
