--- src/server/models/index.js
@@ -1,1 +1,0 @@
-export * from './user';
