+++ src/server/services/index.js
@@ -1,1 +1,2 @@
 export * from './campaigns';
+export * from './users';
